initial alternative test with strategy lists
