+++ client/scripts/routes.js
@@ -4,7 +4,7 @@ var React = require('react/addons');
 var routeActions = require('./actions/routes');
 var IndexPage = React.createFactory(require('./components/index.jsx'));
 
-var StrategiesConsole = React.createFactory(require('./components/admin/strategies.jsx'));
+var ListConsole = React.createFactory(require('./components/admin/adminlist.jsx').strategy_list);
 var ProductsConsole = React.createFactory(require('./components/admin/products.jsx'));
 var LivingSystemsConsole = React.createFactory(require('./components/admin/livingsystems.jsx'));
 var PhenomenaConsole = React.createFactory(require('./components/admin/phenomena.jsx'));
@@ -45,8 +45,8 @@ var index = function() {
     render(IndexPage);
 };
 
-var console_strategies = function() {
-    render(StrategiesConsole);
+var console_list = function() {
+    render(ListConsole);
 };
 
 var console_products = function() {
@@ -152,7 +152,7 @@ var infinite_demo = function() {
 
 var routes = {
   '/': index,
-  '/admin/strategies': console_strategies,
+  '/admin/strategies': console_list,
   '/admin/products': console_products,
   '/admin/living-systems': console_livingsystems,
   '/admin/phenomena': console_phenomena,
